refactor: isolate legacy handled-turn adoption
